requirements.txt
# LivePilotAI - AI智慧導播平台依賴套件

# 核心AI/ML框架
# tensorflow>=2.13.0  # Removed for lightweight deployment
opencv-python>=4.8.0
# torch>=2.0.0       # Removed for lightweight deployment
# torchvision>=0.15.0
numpy>=1.24.0
# pandas>=2.0.0      # Removed if not strictly needed (keep if analyzing CSVs)
# scikit-learn>=1.3.0

# 語音識別
SpeechRecognition>=3.10.0
PyAudio>=0.2.13

# 電腦視覺與影像處理
mediapipe>=0.10.0
# dlib>=19.24.0
# face-recognition>=1.3.0
pillow>=10.0.0

# Web框架與API
fastapi>=0.100.0
uvicorn>=0.23.0
websockets==11.0.2
python-socketio>=5.8.0

# 串流與媒體處理
ffmpeg-python>=0.2.0
av>=10.0.0
imageio>=2.31.0

# 資料庫與快取
redis>=4.6.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# 工具與輔助
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
aiofiles>=23.0.0
python-multipart>=0.0.6
psutil>=5.9.0

# 開發與測試
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 部署與監控
gunicorn>=21.0.0
prometheus-client>=0.17.0
structlog>=23.0.0

# OBS整合
obs-websocket-py>=1.0.0
websockets==11.0.2
websocket-client>=1.6.0
